+++ src/DJTable.h
@@ -49,7 +49,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    crossfadeSlider(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -62,7 +62,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    effectDial(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -75,7 +75,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    stickX(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -88,7 +88,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    stickY(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -101,7 +101,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    rightDJTable(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -114,7 +114,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    leftDJTable(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    const uint16_t myMin = 0;
 	    const uint16_t myZero = 125;
@@ -127,7 +127,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    euphoriaButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -137,7 +137,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    plusButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -147,7 +147,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    minusButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -157,7 +157,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    leftGreenButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -167,7 +167,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    leftRedButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -177,7 +177,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    leftBlueButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -187,7 +187,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    rightGreenButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -197,7 +197,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    rightRedButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
@@ -207,7 +207,7 @@ class DJTable : public Accessory
 	  {
 	    public:
 	    rightBlueButton(uint8_t chan,uint8_t max,uint8_t zero,uint8_t min) : Mapping( chan, max, zero, min) {};
-	    uint16_t mapVar();
+	    unsigned int  mapVar();
 	    void printMap(Stream& stream);
 	    
 	  };
